Fix some other warnings

+++ TownOfUs/Modules/DraftMode/DraftRPCS.cs
@@ -139,7 +139,7 @@ public static class DraftRpcs
         if (slot == localSlot)
         {
             DraftScreenController.Hide();
-            DraftScreenController.ShowFinalPickNotification((ushort)roleId);
+            DraftScreenController.ShowFinalPickNotification(roleId);
         }
     }
